Refactor timer into modules and fix activity sequencing
